// src/interfaces/pations.interface.ts
export interface Pations {
  _id: string;
  firstName: string;
  lastName: string;
  cin: string;
  age: number;
  profession: string;
  address: string;
  phone: string;
  cardiovasculaires: string;
  hematologiques: string;
  endocriniennes: string;
  allergies: string;
  nerveux: string;
  autre: string;
}